com.collabnet.ccf.pi.qc.QCReader:
01/04/08:

-	Added 4 new parameters – targetRepositoryId, targetRepositoryKind, targetSystemId and targetSystemKind. These values will be coming from the HSQL Table and will be used to populate the GenericArtifact. 
-	Modified readModifiedDefects method:
o	Removed the hard-coding of the output file and included this as one of the parameters that can be configured in the spiring wiring file.
o	Added a functionality to store the output (list of genericArtifacts) as an object in an object array that can be used, when stored in an output file, to verify the results in the QCReaderOutputVerificator component.
o	
-	Added 4 new get methods corresponding to each of the new parameters.



git-svn-id: https://ctf.open.collab.net/svn/repos/ccf/trunk@112 6771e1e2-0f46-0410-b5d7-d846ff89f743