Fix for ccf267
The retry mechanism comes into play when the handleException reports that the exception thrown from the components is indeed a connection exception. But when the connection is restored the retry flag is not reset. Consequently the it goes on calling the component's method in an indefinite loop until the connector process itself is restarted.

Fixed this by flipping off the retry flag in case of a smooth communication to the end end system.

git-svn-id: https://ctf.open.collab.net/svn/repos/ccf/trunk@1982 6771e1e2-0f46-0410-b5d7-d846ff89f743