update XSLTs to support the TF5.4 story points field as follows:
SWP2TF: the Backlog Effort field and the points field are both set to the value of the SWP estimate field.

TF2SWP: if the Backlog effort field exists, its value is used for the SWP estimate field. If not, the points field is used. If neither exists, nothing happens.

git-svn-id: https://ctf.open.collab.net/svn/repos/ccf/trunk/core@3826 6771e1e2-0f46-0410-b5d7-d846ff89f743