[artf7205]: Support data type attribute in Rest api
Changes done:

This check in is done to add max attachment size as ccfcoreproperty in ccfcoredefaultconfig.xml.
Since this property is used by both direction the name for the property will be "ccf.direction.max.attachmentsize".
Modified config.xml to have the new property name.
And also modified all the label names with spaces.



git-svn-id: https://ctf.open.collab.net/svn/repos/ccf/trunk/core@5344 6771e1e2-0f46-0410-b5d7-d846ff89f743