[artf7655] Port JIRA implementation to CCF 2.X
Downgrading the JIRATF hsqldb to 1.8 from 2.0

Reason: As CCFMaster 2.2  supports primarily for TFSTF integration which runs only with hsqldb1.8. For demo to accommodate and run  both JIRATF  and TFSTF integration in same CCFMaster we need to use hsqldb1.8 with no choice left. 

Moreover JIRATF is a secondary prototype hence using hsqldb1.8 wont have any issues as of now.

git-svn-id: https://ctf.open.collab.net/svn/repos/ccf/branches/CCF2Jira/core@5522 6771e1e2-0f46-0410-b5d7-d846ff89f743