[artf7748] Obfuscate password for generic participant kind
Although the obfuscate is implemented for generic participant kind in CCF2.2, TFS plugin  which are using the obfuscate password is not added with deobfuscate function. This will make TFS sync fails when the cores runs against CCF 2.2.

Adding the deobfuscate to password property on TFSReader and TFSwriter.

And also remember to add steps in the wiki/words on obfuscation support for generic participant , for those who migrate CCF1.5 plugin to CCF2.x plugin and who wish to run the plugin against CCF2.2 generic platform ,must deobfuscate the password.

 

git-svn-id: https://ctf.open.collab.net/svn/repos/ccf/branches/CCF4TFS/core@5577 6771e1e2-0f46-0410-b5d7-d846ff89f743