Possiblity treat requirement types as one
Changes done:

The main idea behind this story is to sync different types of requirement in QC to a single tracker in TF. To do this one have to create a project mapping for a TF tracker to a QC requirement with type declared as "ALL" (e.g. QCDomain-CCFDemoProjectInQC-ALL => tracker 12).Also sample custom xsl file is checked in, so that developers can use it wisely while defining field mapping.

git-svn-id: https://ctf.open.collab.net/svn/repos/ccf/trunk/core@5733 6771e1e2-0f46-0410-b5d7-d846ff89f743