Updated soap60 support for closeConnection() and isAlive()
On promoting the build to dev env, noticed that TF connections reached
max as configured and gives below error as the earlier soap60
connections are not closed and gives invalid session id as soap50 logoff
is called for the original soap60 login

Could not create connection to the TF system. Max connections reached
for http://teamforge

Now added soap60 support to closeConnection() and isAlive() so that
soap60 calls are logged of when ever closeConnection() is called for
soap60 login.

Change-Id: I3212587e3aaf9f97d525b381dcee65c968a672d2