Support for planning folder point capacity folder
Fixed the minor review comments

Added Points capacity and Display Effort in (trackerunitid) fields for
planning folder create/update.

point capacity field introduced in TF 7.1 (support63())
trackerunitid field introduced in TF 7.0 (support62())

trackerunitid cannot be set to null (None). From hcn Default base unit
is Hours.
http://help.collab.net/index.jsp?topic=/teamforge710/action/trackers-configuringtrackerunits.html

I'm not able to find any TF SOAP call that provides default base
unit.Also on creating planning folder through webui, by default ; base
unit is set as Hours.

Hence in the code if the trackerunitid is null or when not mapped, the
default base unit Hours is set to the field. If the corresponding
trackerunitid field in QC is mapped with a value (say Hours or Week), TF
soap call is made to find the technical id for the tracker unit id (say
for ex Hours <-> tknt1001) and that is set back to TF.

Tested against TF7.1 , able to ship the point capacity and trackerunitid
field values bi-directionally.

Change-Id: I1677c257c23064ef9c25f8037bed414b945094c0