Support for ALM 12
With HP ALM 12 introducing web client and desktop client.The comments
added in one client get auto synched to another. But the display tags
generated for comments differ from web and desktop client which makes
the current comment parsing logic broken. 
For  each comment added from HP ALM web client and desktop client the
generated html tags are getting differed . And when every time the
comments are added ; the old comments gets altered with respect to the
clients and some time the old value length is lesser than the new value
length as some html tags are deleted and modified. This makes the
comment 
parsing difficult.

Hence new comment parsing logic to strip all the html tags out of the
comments and extract only the values out of it. This new comment parsing
logic is applicable only for HP ALM 12. Earlier comments parsing for HP
ALM 11 and previous versions are left  untouched.

Externalized lot of html values to constants on both new code as well as
old code.

Refactored the call to qcc.getMajorVersion() such that its doesnt repeat
the call to QC backend for every check.


Change-Id: I9f217ccbd16e2dfcc7325a778758014ad3111c64