RMD Simulation - Dryrun Mode feature
Introduced new dry run mode values (after/beforeTransformation,stop)
When dryrun set to stop, AbstractReader will clear the artifacts in
pipeline and will set the dry-run to off - implicitly.

For the selected RMD , set the RMD configuration (name -dryrun values
-beforeTransformation, afterTransformation,stop) to work with
dryrun.

If the dryrun set to before/afterTransformation for the selected RMD,
the artifacts are stored in hospital for the admin to preview and
validate the workitem synchronization and the synchronization table is
not updated when dry-run set to before/afterTransformation.

To get out of dryrun mode , the dryrun rmdconfig is set to
stop.AbstractReader inturn sets them to off and the failed shipments
(dryrun) for the given RMD is deleted.
(deletion of corresponding dry run failed shipments for the selected RMD
will be handled via GUI)

Update query to update repositorymappingdirectionconfig table to set the
value to off when dryrun mode is set to stop (sql.prop,config2.xml,
AbstractReader)

Added the info message in EntityService2.java when dryrun mode is on; to
let the user how to reprocess the existing artifacts in hospital

Bumped the ccfcore version to 2.4.0 in build.xml

DryModeHandler abstract class.

NoOpDryModeHandler  - Dummy implementation

RMDDryModeHandler  - Original implementation

config2.xml will have the original implementation(RMDDryModeHandler)
wired in. And the components that use the DryModeHandler ; instantiate
the dummy one by default.
config.xmls wired with original implementation will reassign this
variable in the property setter method.
So incase of different plugin / CCF 1.5.x / SWP, the Core will use the
default impl. By this implementation unnecessary if checks are avoided
and code is more readable.

Moved the dryrun check condition before the DBcall in EntityService2
AbstractReader

Followed naming convention in query. rmdc changed to
repositoryMappingDirectionConfig in sql.prop and AReader , config2.xml

Removed the condition in validate() that breaks the  compatibilty 
Moved the null check condition to updatemethod

Tested dry-run mode with both plainartifact and attachments
bi-directionally


Change-Id: I6ad00dcf549287845b778a31749be4d2ea120228