[artf11056] user tag missing logged in ccf info logs
The comment in ALM is formed in the below format

ccfsync <ccfsync>, 2014-07-09: Second comment

Comment parsing algorithm used for HP ALM 12 for CCF 2.3.1 use jerico
utils to extract the text from HTML and output of 
this conversion doesnt escape/converts the < > in the full name to </>
And during transformation , the whitelist 
funciton calls strip html out of the comment text and hence the comment
synched to TF skips the full name part from the 
comment text and that is reason for the corresponding entry in log.

The reason why we havent got this issue in CCF2.3.1 is the ALM user name
created doesnt have full named added to it.

Used StringUtils.encodeHTMLtoentityreference to escape the html <> so
that now the <fullname> is part of the comment.

Change-Id: I123bc238a4419572ced57597fc7191f5ecfc07e4