Dont rely on prefix for TFAttachmentHandler
Introduced new property ignoreSyncUserNameFromAttachment with
default value set to false. When the ignoreSyncUserNameFromAttachment
property set to false attachment filename is prefixed with sync
username. When the property ignoreSyncUserNameFromAttachment is set to
true sync username will not be prefixed on the attachment filename. This
change applicable only for QC2TF direction since TF2QC attachment
filename is not prefixed.

* Fixed the removal of whitespace in the empty comment line by setting
up the CCF Code formatter

Change-Id: I1b070ea8668ed9d881715df8d4fb8eb928f6b166