Ignore requirement root (requirement) folder
Added a new property ignorerequirementrootfolder in qc2tf direction with
the default value set as false. Requirement root folder (0) sync skipped
between qc2tf if the ignorerequirementrootfolder property set to true.

1. QCReader.java -> Included getter&setter for the above property and
setting the value for parentId as unknown when id 0.
2. QCHandler.java -> Updated the sqlQuery method to skip the root folder
3. ccfcoredefaultconfig.xml -> Added a new property
ignorerequirementrootfolder to expose it in ccfmaster.
4. config2.xml -> Added the property ignorerequirementrootfolder and set
the defalut value as false.

Change-Id: I11e7842fe62c5a197822ced3f31a49091c4abcad