Fix comment handling issues with HP ALM 12.20
Modified getDeltaOfComment and cleanUpComment methods in QCHandler
to properly handle HP ALM 12.2 comment format.
1. In getDeltaOfComment, all characters before first '\r' will be
skipped. This is to get the new comment values alone and omit
the extra characters carry forwarded when switching between clients.

2. In cleanUpComment, remove the initial underscore string.
New comment now start with an underscore string, so removing it
to get the comment value. Also removing <br> tags in case we are
dealing with HP ALM 11. 

Change-Id: I051fc68091a8978512f699296540f8e9161b2002