Handle updates of QC artifacts that got audit history removed
This is to handle updates of QC artifacts from TF that got audit_log
history removed. So it will take '0' as targetTransactionIdBeforeUpdate
instead of null, in case of audit_log has removed before to the updates
of QC artifact. 

Added a new getTargetTransactionIdBeforeUpdate method, which will return
the targetTransactionIdBeforeUpdate to both updateDefect and
updateRequirement. This method will return 0 in case audit history has
been removed.

Change-Id: I7cf88684409938ab6de78bc10ae143ec8b96b5e8